move api clients to own package
