+++ src/app/src/test/kotlin/org/imperial/mrc/hint/unit/controllers/MetadataControllerTests.kt
@@ -3,7 +3,7 @@ package org.imperial.mrc.hint.unit.controllers
 import com.nhaarman.mockito_kotlin.doReturn
 import com.nhaarman.mockito_kotlin.mock
 import org.assertj.core.api.Assertions.assertThat
-import org.imperial.mrc.hint.HintrAPIClient
+import org.imperial.mrc.hint.clients.HintrAPIClient
 import org.imperial.mrc.hint.controllers.MetadataController
 import org.junit.jupiter.api.Test
 import org.springframework.http.ResponseEntity
